Add dynamics_on_manifolds scala project with IntelliJ IDE
