+++ build.sbt
@@ -1,0 +1,8 @@
+name := "dynamics_on_manifolds"
+
+version := "0.1"
+
+scalaVersion := "2.13.2"
+
+libraryDependencies += "org.scala-lang.modules" %% "scala-parser-combinators" % "1.1.2"
+libraryDependencies += "org.scalatest" %% "scalatest" % "3.0.8" % Test
