--- src/MatrixTypes.scala
@@ -1,10 +1,0 @@
-// MatrixTypes.scala
-package dynamics_on_manifolds
-
-trait SymmetricMatrix {}
-
-trait ConstantMatrix {}
-
-trait SkewSymmetricMatrix {}
-
-trait Unimodular {}
