--- src/VectorTypes.scala
@@ -1,7 +1,0 @@
-// VectorTypes.scala
-package dynamics_on_manifolds
-
-trait UnitVector {}
-
-trait ConstantVector {}
-
