+++ src/main/scala/hybridrobotics/dynamics/Main.scala
@@ -1,0 +1,7 @@
+package hybridrobotics.dynamics
+
+import hybridrobotics.dynamics.mechanical_systems.SphericalPendulum
+
+object Main extends App {
+  SphericalPendulum.main()
+}
