+++ src/main/scala/hybridrobotics/dynamics/operations/MatrixTypes.scala
@@ -1,0 +1,10 @@
+// MatrixTypes.scala
+package hybridrobotics.dynamics.operations
+
+trait SymmetricMatrix {}
+
+trait ConstantMatrix {}
+
+trait SkewSymmetricMatrix {}
+
+trait Unimodular {}
