+++ src/main/scala/hybridrobotics/dynamics/operations/VectorTypes.scala
@@ -1,0 +1,7 @@
+// VectorTypes.scala
+package hybridrobotics.dynamics.operations
+
+trait UnitVector {}
+
+trait ConstantVector {}
+
